# shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API de la app de pulperías",
    "version": "1.0.0",
    "description": "Contrato escrito a mano (D-17). Los clientes HTTP del móvil y de la web se escriben contra este archivo, y una prueba de la API comprueba que las respuestas reales lo cumplen. Dinero: entero en la unidad menor (centavos de lempira). Cantidades: enteros en milésimas. Fechas: ISO 8601 en UTC. Todo objeto declara todas sus propiedades: una propiedad que no esté aquí no existe. Las rutas de un negocio exigen el token de acceso y la cabecera X-Business-Id; sin la cabecera responden 400 business_required y si el usuario no es miembro activo, 403 forbidden. Los errores devuelven un código estable (RNF-5) que los clientes traducen al español."
  },
  "servers": [{ "url": "http://localhost:5109" }],
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/api/businesses": {
      "get": {
        "summary": "Negocios del usuario con su rol en cada uno (RF-5, RF-6).",
        "responses": {
          "200": { "description": "Lista de negocios.", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/BusinessSummary" } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "summary": "Crear un negocio adicional; quien lo crea es su dueño (RF-79).",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateBusinessRequest" } } } },
        "responses": {
          "201": { "description": "Negocio creado.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BusinessSummary" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/admin/businesses": {
      "get": {
        "summary": "Negocios de la plataforma con sus cifras agregadas, por nombre (RF-97, D-29). Solo super administradores; nunca trae datos de fiados.",
        "parameters": [
          { "name": "page", "in": "query", "required": false, "description": "Página desde 1; por omisión 1.", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "pageSize", "in": "query", "required": false, "description": "Por omisión 25, máximo 100.", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "name": "search", "in": "query", "required": false, "description": "Parte del nombre del negocio o del correo de un dueño.", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/BusinessStatus" } }
        ],
        "responses": {
          "200": { "description": "Una página de negocios.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminBusinessPage" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/admin/businesses/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": {
        "summary": "La ficha de un negocio: dueños, miembros, estado, última sincronización y conteos (RF-97, D-29).",
        "responses": {
          "200": { "description": "El negocio.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminBusiness" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/admin/businesses/{id}/suspend": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "summary": "Suspender un negocio con un motivo (RF-98, D-29). Conserva sus datos y rechaza sus peticiones, lotes y pull con business_suspended hasta reactivarse.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReasonRequest" } } } },
        "responses": {
          "200": { "description": "El negocio, ya suspendido.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminBusiness" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/admin/businesses/{id}/reactivate": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "summary": "Reactivar un negocio suspendido (RF-98, D-29). Todo vuelve a funcionar y la cola de los teléfonos se aplica.",
        "responses": {
          "200": { "description": "El negocio, ya activo.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminBusiness" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/admin/businesses/{id}/activate": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "summary": "Activar un negocio pendiente (RF-103, D-30). Rechazarlo es suspenderlo con un motivo.",
        "responses": {
          "200": { "description": "El negocio, ya activo.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminBusiness" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/admin/accounts/{id}/suspend": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "summary": "Suspender una cuenta con un motivo (RF-99, D-29): no inicia sesión (account_suspended) y sus sesiones se cierran; sus negocios siguen intactos. Un super administrador no puede suspenderse a sí mismo.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReasonRequest" } } } },
        "responses": {
          "200": { "description": "La cuenta, ya suspendida.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminAccount" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/admin/accounts/{id}/reactivate": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "summary": "Reactivar una cuenta suspendida (RF-99, D-29).",
        "responses": {
          "200": { "description": "La cuenta, ya activa.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminAccount" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/admin/accounts/{id}/reset-password": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "summary": "Restablecer la contraseña de una cuenta (RF-100, D-29): genera una nueva que se muestra una sola vez y cierra las sesiones de la cuenta. No incluye datos de negocios.",
        "responses": {
          "200": { "description": "La contraseña nueva, para entregársela a la persona.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewPassword" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/admin/audit": {
      "get": {
        "summary": "La auditoría de la plataforma, de la más reciente a la más antigua (RF-101, D-29): quién, cuándo, qué, sobre qué y el motivo. Solo lectura.",
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1 } },
          { "name": "pageSize", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "name": "accountId", "in": "query", "required": false, "description": "Solo lo hecho sobre esa cuenta.", "schema": { "type": "string", "format": "uuid" } },
          { "name": "businessId", "in": "query", "required": false, "description": "Solo lo hecho sobre ese negocio.", "schema": { "type": "string", "format": "uuid" } },
          { "name": "action", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/AdminAction" } }
        ],
        "responses": {
          "200": { "description": "Una página de la auditoría.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminAuditPage" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/admin/accounts": {
      "get": {
        "summary": "Cuentas de la plataforma por correo, con su estado y sus negocios (RF-97, D-29). Solo super administradores.",
        "parameters": [
          { "name": "page", "in": "query", "required": false, "description": "Página desde 1; por omisión 1.", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "pageSize", "in": "query", "required": false, "description": "Por omisión 25, máximo 100.", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "name": "search", "in": "query", "required": false, "description": "Parte del correo.", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Una página de cuentas.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminAccountPage" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/admin/accounts/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": {
        "summary": "La ficha de una cuenta (RF-97, D-29).",
        "responses": {
          "200": { "description": "La cuenta.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminAccount" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "bearer": { "type": "http", "scheme": "bearer", "description": "Token de acceso opaco, de 15 minutos (D-27)." } },
    "parameters": {
...
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
    },
    "responses": {
      "BadRequest": { "description": "Petición mal formada o rechazada por una regla (código estable).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Falta la sesión, caducó o las credenciales no sirven.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "El usuario no pertenece al negocio o su rol no permite la acción.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "No existe en el negocio o el usuario no puede verlo.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "La acción choca con el estado actual.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["code", "codes"],
        "properties": {
          "code": { "type": "string", "description": "El código principal, estable." },
          "codes": { "type": "array", "items": { "type": "string" }, "description": "Todos los problemas encontrados; el primero es code." }
        }
      },
      "Role": { "type": "string", "enum": ["owner", "employee"] },
      "AmountMode": { "type": "string", "enum": ["integer", "two_decimals"] },
      "QuantityMode": { "type": "string", "enum": ["integer", "fractional"] },
...
      "BusinessStatus": { "type": "string", "enum": ["pending", "active", "suspended"] },
      "AdminBusiness": {
        "type": "object",
        "description": "Metadatos y cifras agregadas de un negocio; nunca nombres de clientes ni de productos, montos ni deudas (RF-97).",
        "required": ["id", "name", "ownerEmails", "memberCount", "createdAt", "status", "statusReason", "lastSyncAt", "clientCount", "productCount", "fiadoCount", "paymentCount"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "ownerEmails": { "type": "array", "items": { "type": "string" } },
          "memberCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "status": { "$ref": "#/components/schemas/BusinessStatus" },
          "statusReason": { "type": "string", "nullable": true },
          "lastSyncAt": { "type": "string", "format": "date-time", "nullable": true },
          "clientCount": { "type": "integer" },
          "productCount": { "type": "integer" },
          "fiadoCount": { "type": "integer" },
          "paymentCount": { "type": "integer" }
        }
      },
      "AdminAccountBusiness": {
        "type": "object",
        "required": ["id", "name", "role", "status"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "status": { "$ref": "#/components/schemas/BusinessStatus" }
        }
      },
      "AdminAccount": {
        "type": "object",
        "required": ["id", "email", "createdAt", "isSuperAdmin", "suspendedAt", "suspensionReason", "businesses"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "isSuperAdmin": { "type": "boolean" },
          "suspendedAt": { "type": "string", "format": "date-time", "nullable": true },
          "suspensionReason": { "type": "string", "nullable": true },
          "businesses": { "type": "array", "items": { "$ref": "#/components/schemas/AdminAccountBusiness" } }
        }
      },
      "ReasonRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": { "reason": { "type": "string", "description": "Obligatorio y no vacío: queda en la auditoría." } }
      },
      "NewPassword": {
        "type": "object",
        "required": ["password"],
        "properties": { "password": { "type": "string", "description": "16 caracteres; no se vuelve a mostrar." } }
      },
      "AdminAction": {
        "type": "string",
        "enum": ["reset_password", "grant_superadmin", "revoke_superadmin", "suspend_business", "reactivate_business", "activate_business", "suspend_account", "reactivate_account"]
      },
      "AdminAuditEntry": {
        "type": "object",
        "required": ["id", "action", "performedBy", "performedAt", "targetUserId", "targetEmail", "targetBusinessId", "targetBusinessName", "detail"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "action": { "$ref": "#/components/schemas/AdminAction" },
          "performedBy": { "type": "string", "description": "El correo del super administrador o el operador del comando del servidor." },
          "performedAt": { "type": "string", "format": "date-time" },
          "targetUserId": { "type": "string", "format": "uuid", "nullable": true },
          "targetEmail": { "type": "string", "nullable": true },
          "targetBusinessId": { "type": "string", "format": "uuid", "nullable": true },
          "targetBusinessName": { "type": "string", "nullable": true },
          "detail": { "type": "string", "nullable": true, "description": "El motivo, cuando la acción lo lleva." }
        }
      },
      "AdminAuditPage": {
        "type": "object",
        "required": ["items", "page", "pageSize", "total"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/AdminAuditEntry" } },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "AdminBusinessPage": {
        "type": "object",
        "required": ["items", "page", "pageSize", "total"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/AdminBusiness" } },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "AdminAccountPage": {
        "type": "object",
        "required": ["items", "page", "pageSize", "total"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/AdminAccount" } },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
...
      "CreateBusinessRequest": {
        "type": "object",
        "required": ["name", "amountMode", "quantityMode"],
        "properties": {
          "name": { "type": "string" },
          "amountMode": { "$ref": "#/components/schemas/AmountMode" },
          "quantityMode": { "$ref": "#/components/schemas/QuantityMode" }
        }
      },
      "BusinessSummary": {
        "type": "object",
        "required": ["id", "name", "role", "amountMode", "quantityMode", "status"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "amountMode": { "$ref": "#/components/schemas/AmountMode" },
          "quantityMode": { "$ref": "#/components/schemas/QuantityMode" },
          "status": { "$ref": "#/components/schemas/BusinessStatus" }
        }
      },
...
    }
  }
}